requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
SQLAlchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
httpx>=0.27.0
psycopg2-binary>=2.9.9
Jinja2>=3.1.0
python-multipart>=0.0.9
beautifulsoup4>=4.12.0
slowapi>=0.1.9
